+++ ScoutSuite/providers/aws/configs/regions_config.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from ScoutSuite.providers.base.configs.resources import Resources
-from ScoutSuite.providers.aws.aws_facade import AWSFacade
+from ScoutSuite.providers.aws.facade import AWSFacade
 
 
 class RegionsConfig(Resources):
