+++ ScoutSuite/providers/aws/facade.py
